pyproject.toml
[project]
name = "poshub-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Cinex10",email = "y.driss@esi-sba.dz"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.13,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "poshub_api", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
asgi-lifespan = "^2.1.0"
coverage = "^7.9.1"
trio = "^0.30.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
